pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "league-analysis-mcp-server"
version = "0.1.0"
description = "Model Context Protocol server for Yahoo Fantasy Sports API with advanced historical analysis and manager profiling"
authors = [{name = "League Analysis MCP", email = "info@league-analysis-mcp.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
keywords = [
    "mcp",
    "model-context-protocol", 
    "fantasy-sports",
    "yahoo-fantasy",
    "nfl",
    "nba",
    "mlb",
    "nhl",
    "analytics",
    "ai-assistant"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Games/Entertainment",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Framework :: AsyncIO",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastmcp>=2.0.0",
    "yfpy>=16.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "asyncio-throttle>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/league-analysis-mcp/league-analysis-mcp"
Repository = "https://github.com/league-analysis-mcp/league-analysis-mcp.git"
Issues = "https://github.com/league-analysis-mcp/league-analysis-mcp/issues"
Documentation = "https://github.com/league-analysis-mcp/league-analysis-mcp#readme"
Changelog = "https://github.com/league-analysis-mcp/league-analysis-mcp/blob/main/CHANGELOG.md"

[project.scripts]
league-analysis-mcp-server = "league_analysis_mcp_server.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/league_analysis_mcp_server"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/config", 
    "/README.md",
    "/LICENSE",
    "/CHANGELOG.md"
]